refactor(igxOverlay): create ScrollStrategy abstract class
TypeDocs may take dev comments only from interfaces when @inheritdoc is added.
However, scroll strategies need to have constructor. This is why we add
ScrollStrategy abstract class. It is implementing IScrollStrategy, and then
all implementations extends the new abstract class
