+++ projects/igniteui-angular/src/lib/services/overlay/position/global-position-strategy.ts
@@ -12,15 +12,15 @@ export class GlobalPositionStrategy implements IPositionStrategy {
         closeAnimation: fadeOut
     };
 
+    /** @inheritdoc */
     public settings: PositionSettings;
+
     constructor(settings?: PositionSettings) {
         this.settings = Object.assign({}, this._defaultSettings, settings);
     }
 
-    /**
-     * @inheritdoc
-     */
-    position(contentElement: HTMLElement, size?: { width: number, height: number}, document?: Document, initialCall?: boolean): void {
+    /** @inheritdoc */
+    position(contentElement: HTMLElement, size?: { width: number, height: number }, document?: Document, initialCall?: boolean): void {
         switch (this.settings.horizontalDirection) {
             case HorizontalAlignment.Left:
                 contentElement.parentElement.style.justifyContent = 'flex-start';
