+++ projects/igniteui-angular/src/lib/services/overlay/scroll/index.ts
@@ -1,7 +1,9 @@
+
 // Export scroll strategies
+export * from './scroll-strategy';
+export * from './IScrollStrategy';
 export * from './absolute-scroll-strategy';
 export * from './block-scroll-strategy';
 export * from './close-scroll-strategy';
-export * from './IScrollStrategy';
 export * from './NoOpScrollStrategy';
